+++ examples/country/lib/ui/screen/country_list_screen/country_list_screen_widget_model.dart
@@ -4,6 +4,8 @@ import 'package:country/ui/screen/country_list_screen/country_list_screen.dart';
 import 'package:country/ui/screen/country_list_screen/country_list_screen_model.dart';
 import 'package:dio/dio.dart';
 import 'package:elementary/elementary.dart';
+import 'package:elementary_helper/elementary_helper.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -24,7 +26,7 @@ class CountryListScreenWidgetModel
   final _countryListState = EntityStateNotifier<List<Country>>();
 
   @override
-  ListenableState<EntityState<List<Country>>> get countryListState =>
+  ValueListenable<EntityState<List<Country>>> get countryListState =>
       _countryListState;
 
   @override
@@ -50,16 +52,16 @@ class CountryListScreenWidgetModel
   void onErrorHandle(Object error) {
     super.onErrorHandle(error);
 
-    if (error is DioError &&
-        (error.type == DioErrorType.connectTimeout ||
-            error.type == DioErrorType.receiveTimeout)) {
+    if (error is DioException &&
+        (error.type == DioExceptionType.connectionTimeout ||
+            error.type == DioExceptionType.receiveTimeout)) {
       ScaffoldMessenger.of(context)
           .showSnackBar(const SnackBar(content: Text('Connection troubles')));
     }
   }
 
   Future<void> _loadCountryList() async {
-    final previousData = _countryListState.value?.data;
+    final previousData = _countryListState.value.data;
     _countryListState.loading(previousData);
 
     try {
@@ -72,8 +74,8 @@ class CountryListScreenWidgetModel
 }
 
 /// Interface of [CountryListScreenWidgetModel]
-abstract class ICountryListWidgetModel extends IWidgetModel {
-  ListenableState<EntityState<List<Country>>> get countryListState;
+abstract interface class ICountryListWidgetModel implements IWidgetModel {
+  ValueListenable<EntityState<List<Country>>> get countryListState;
 
   TextStyle get countryNameStyle;
 }