+++ examples/profile/lib/features/profile/screens/interests_screen/interests_screen.dart
@@ -1,4 +1,5 @@
 import 'package:elementary/elementary.dart';
+import 'package:elementary_helper/elementary_helper.dart';
 import 'package:flutter/material.dart';
 import 'package:profile/assets/colors/colors.dart';
 import 'package:profile/assets/strings/interests_screen_strings.dart';
@@ -53,7 +54,7 @@ class InterestsScreen extends ElementaryWidget<IInterestsScreenWidgetModel> {
 class _ContentWidget extends StatelessWidget {
   final List<String>? listInterests;
   final ListenableState<List<String>> listUserInterestsState;
-  final Function({required String interest}) onChanged;
+  final void Function({required String interest}) onChanged;
 
   const _ContentWidget({
     required this.listInterests,
@@ -92,7 +93,7 @@ class _ContentWidget extends StatelessWidget {
 class _ItemInterest extends StatelessWidget {
   final String interest;
   final ListenableState<List<String>> state;
-  final Function({required String interest}) onChanged;
+  final void Function({required String interest}) onChanged;
 
   const _ItemInterest({
     required this.interest,