+++ packages/elementary/example/lib/impl/screen/test_page_widget_model.dart
@@ -4,6 +4,8 @@ import 'package:counter/impl/screen/test_page_model.dart';
 import 'package:counter/impl/screen/test_page_widget.dart';
 import 'package:counter/main.dart';
 import 'package:elementary/elementary.dart';
+import 'package:elementary_helper/elementary_helper.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -17,7 +19,7 @@ TestPageWidgetModel testPageWidgetModelFactory(BuildContext context) {
 class TestPageWidgetModel extends WidgetModel<TestPageWidget, TestPageModel>
     implements ITestPageWidgetModel {
   @override
-  ListenableState<EntityState<int>> get valueState => _valueController;
+  ValueListenable<EntityState<int>> get valueState => _valueController;
 
   late EntityStateNotifier<int> _valueController;
 
@@ -47,8 +49,8 @@ class TestPageWidgetModel extends WidgetModel<TestPageWidget, TestPageModel>
   }
 }
 
-abstract class ITestPageWidgetModel extends IWidgetModel {
-  ListenableState<EntityState<int>> get valueState;
+abstract interface class ITestPageWidgetModel implements IWidgetModel {
+  ValueListenable<EntityState<int>> get valueState;
 
   Future<void> increment();
 }