+++ packages/elementary_test/example/lib/impl/screen/test_page_widget_model.dart
@@ -3,6 +3,8 @@ import 'dart:async';
 import 'package:counter/impl/screen/test_page_model.dart';
 import 'package:counter/impl/screen/test_page_widget.dart';
 import 'package:elementary/elementary.dart';
+import 'package:elementary_helper/elementary_helper.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -17,7 +19,7 @@ class TestPageWidgetModel extends WidgetModel<TestPageWidget, TestPageModel>
   final ThemeWrapper _themeWrapper;
 
   @override
-  ListenableState<EntityState<int>> get valueState => _valueController;
+  ValueListenable<EntityState<int>> get valueState => _valueController;
 
   @override
   TextStyle get counterStyle => _counterStyle;
@@ -62,8 +64,8 @@ class TestPageWidgetModel extends WidgetModel<TestPageWidget, TestPageModel>
   }
 }
 
-abstract class ITestPageWidgetModel extends IWidgetModel {
-  ListenableState<EntityState<int>> get valueState;
+abstract interface class ITestPageWidgetModel implements IWidgetModel {
+  ValueListenable<EntityState<int>> get valueState;
 
   TextStyle get counterStyle;