+++ packages/elementary_test/example/test/country_list_screen_wm_test.dart
@@ -6,6 +6,7 @@ import 'package:counter/impl/screen/test_page_model.dart';
 import 'package:counter/impl/screen/test_page_widget.dart';
 import 'package:counter/impl/screen/test_page_widget_model.dart';
 import 'package:elementary/elementary.dart';
+import 'package:elementary_helper/elementary_helper.dart';
 import 'package:elementary_test/elementary_test.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -75,7 +76,7 @@ void main() {
       final value = wm.valueState.value;
 
       expect(value, isNotNull);
-      expect(value!.isLoading, isTrue);
+      expect(value.isLoadingState, isTrue);
     },
   );
 }