+++ packages/elementary_test/lib/src/widget_model_test.dart
@@ -1,6 +1,6 @@
-// ignore_for_file: avoid_implementing_value_types, implementation_imports, invalid_use_of_visible_for_testing_member, invalid_use_of_protected_member
+// ignore_for_file: invalid_use_of_visible_for_testing_member, invalid_use_of_protected_member
 
-import 'package:elementary/src/core.dart';
+import 'package:elementary/elementary.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -96,9 +96,7 @@ abstract class WMTester<WM extends WidgetModel, W extends ElementaryWidget> {
 }
 
 class _WMTestableElement<WM extends WidgetModel, W extends ElementaryWidget>
-    extends WMContext
-    with Diagnosticable
-    implements Elementary, WMTester<WM, W> {
+    extends WMContext with Diagnosticable implements WMTester<WM, W> {
   final WM _wm;
 
   _WMTestableElement(this._wm);