+++ packages/elementary_test/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 2.0.0
+* supported `elementary` version increased to `^3.0.0`;
+
+## 2.0.0-dev
+### Changed
+* supported `elementary` version increased to `^3.0.0-dev`;
+
 ## 1.1.3
 ### Changed
 * supported `elementary` version increased to `^2.0.0`;